backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-mongoengine==1.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.1
Werkzeug==3.0.1
certifi==2024.2.2
pymongo==4.6.2
dnspython==2.7.0
email-validator==2.2.0
mongoengine==0.29.1
WTForms==3.2.1
Flask-WTF==1.2.2
